+++ code_translated_to_cpp/cpuclock.h
@@ -18,13 +18,13 @@ public:
 	std::string path_;
     int         cpu_id_;
     int         dev_count_; // The number of CPU cores
-	int         cpu_freq_;
-	int         cpu_util_;
+	long long   cpu_freq_;
+	double      cpu_util_;
 
     /**
 	 * Creates a CpuClock object with a path_ of "/sys/devices/system/cpu",
 	 * a cpu_id_ equal to set_cpu_id, a dev_count_ of 8, a cpu_freq_ of 0,
-     * and a cpu_util_ of 0.
+     * and a cpu_util_ of 0.0
      * 
      * @param set_cpu_id Set cpu_id_ to this value.
 	 *
@@ -36,7 +36,7 @@ public:
 	 *
 	 * @return The current CPU frequency
 	 */
-    int GetClock();
+    long long GetClock();
 
 	/**
 	 * Sets this CPU's frequency
@@ -44,9 +44,9 @@ public:
 	 * @param new_freq We set the CPU frequency to this value and store it in cpu_freq_
 	 *				   
 	 */
-    void SetClock(int new_freq);
+    void SetClock(long long new_freq);
 
-	std::vector<int> GetAllClock(); // Not implemented
+	std::vector<long long> GetAllClock(); // Not implemented
 	
-	int GetUtilization(); // Not implemented
+	double GetUtilization(); // Not implemented
 };
